move NodeClient trait impl with struct + remove NodeClientExt
